// src/types/Iconfigure.ts
export interface Iconfigure {
    telephone: string;
    telephoneValidate:string,
    operation: string;
    minuts: string;
    internet: string;
    rentWiFiRouter: boolean;
    ransomWiFiRouter: boolean;
    faceBook: boolean;
    vK: boolean;
    oK: boolean;
    instagram: boolean;
    tikTok: boolean;
  }